feat(pokemon): DJ-30 add controller for updating pokemon treinador
This commit introduces the `UpdatePokemonTreinadorController` which
facilitates the process of updating the associated "treinador"
(Trainer) of a specified Pokemon. The primary motivation for this
change is to improve the functionality of the Pokemon management
API by allowing trainers to be updated dynamically, which is a
common requirement in managing Pokemon relationships.

The new controller includes validation for the request data to ensure
that both the Pokemon ID and the trainer details are provided.
Additionally, the controller interacts with the `PokemonRepository`
to perform the update operation and returns an HTTP 204 No Content
response upon a successful update.

Furthermore, this commit includes the necessary updates to the
routing configuration to accommodate the new endpoint, allowing
an HTTP PUT request to `/pokemons/:id` that utilizes the newly
implemented controller. This feature aims to enhance the overall
user experience by providing better data management capabilities
within the application.