feat(pokemons): DJ-30 add endpoint to load Pokemon by ID
This commit introduces a new endpoint to the existing Express.js
router, allowing users to retrieve a specific Pokemon's details
by its ID. The implemented feature uses the `LoadPokemonByIdController`
to fetch data and respond to GET requests made to the route
`/pokemons/:id`.

The addition of this endpoint is essential for enhancing
the API's capabilities as it provides a way for clients to
access detailed information about a specific Pokemon,
thus improving the overall user experience. The controller
is instantiated using a factory function, `makeLoadPokemonByIdController`,
which encapsulates the repository logic, keeping the codebase
consistent with the current architecture.

By enabling this functionality, we aim to fulfill the
requirements laid out in ticket DJ-30, contributing to
the CRUD operations of the Pokemon management feature.