feat(pokemon): DJ-30 Add delete pokemon controller
This commit introduces a new controller for deleting Pokémon
from the system. The `DeletePokemonController` implements
necessary request validation and deletion logic.

The controller checks that the provided ID is present and valid
before attempting to delete the Pokémon. If the ID is invalid
or not a number, a validation error is returned. Additionally,
if the Pokémon to be deleted is not found, a 404 response is
returned.

This functionality is essential for managing Pokémon records,
allowing users to remove specific Pokémon from their collection.
Adding this controller aligns with our goal to implement a full
CRUD (Create, Read, Update, Delete) API for Pokémon management
as defined in ticket DJ-30.